Num - 3 is wrong, instead it should just be num.
The previous solution doesn't work on the FCC console, it does not pass 2 tests. 

According to the question, the slice should start at 0 and end at num ( not num-3). And then concatenate '...' if and only if the string length is greater than the num value passed as the second argument. 

num-3 is nowhere mentioned in the question and it is totally wrong.
 
Changed both basic solution ( with IF statement ) and advanced solution with (ternary operator).